Diff viewer, labels management and branches

+++ pages/[userId]/[repositoryId]/issues/[issueIid].js
@@ -17,13 +17,16 @@ import RepositoryMainTabs from "../../../../components/repository/mainTabs";
 import Footer from "../../../../components/footer";
 import CommentEditor from "../../../../components/repository/commentEditor";
 import CommentView from "../../../../components/repository/commentView";
+import SystemCommentView from "../../../../components/repository/systemCommentView";
 import {
   deleteComment,
-  updateIssue,
+  updateIssueLabels,
+  updateIssueAssignees,
 } from "../../../../store/actions/repository";
 import AssigneeSelector from "../../../../components/repository/assigneeSelector";
 import LabelSelector from "../../../../components/repository/labelSelector";
 import getIssueAllLabels from "../../../../helpers/getIssueAllLabels";
+import Label from "../../../../components/repository/label";
 
 export async function getServerSideProps() {
   return { props: {} };
@@ -57,10 +60,10 @@ function RepositoryIssueView(props) {
         router.query.issueIid
       ),
     ]);
-    const al = await getIssueAllLabels(r.id);
     if (r) setRepository(r);
     if (i) setIssue(i);
-    setAllLabels(al);
+    setAllLabels(r.labels);
+    console.log(i);
   }, [router.query]);
 
   const getAllComments = async () => {
@@ -200,31 +203,40 @@ function RepositoryIssueView(props) {
                       </div>
                     </div>
                     <div className="p-4 markdown-body">
-                      <ReactMarkdown>{issue.description}</ReactMarkdown>
+                      <ReactMarkdown linkTarget="_blank">
+                        {issue.description}
+                      </ReactMarkdown>
                     </div>
                   </div>
                 </div>
                 {allComments.map((c, i) => {
-                  return (
-                    <CommentView
-                      comment={c}
-                      userAddress={props.selectedAddress}
-                      onUpdate={async (id) => {
-                        const newComment = await getComment(id);
-                        const newAllComments = [...allComments];
-                        newAllComments[i] = newComment;
-                        setAllComments(newAllComments);
-                      }}
-                      onDelete={async (id) => {
-                        const res = await props.deleteComment({ id });
-                        if (res && res.code === 0) {
+                  if (c.system) {
+                    return (
+                      <SystemCommentView comment={c} key={"comment" + i} />
+                    );
+                  } else {
+                    return (
+                      <CommentView
+                        comment={c}
+                        key={"comment" + i}
+                        userAddress={props.selectedAddress}
+                        onUpdate={async (id) => {
+                          const newComment = await getComment(id);
                           const newAllComments = [...allComments];
-                          newAllComments.splice(i, 1);
+                          newAllComments[i] = newComment;
                           setAllComments(newAllComments);
-                        }
-                      }}
-                    />
-                  );
+                        }}
+                        onDelete={async (id) => {
+                          const res = await props.deleteComment({ id });
+                          if (res && res.code === 0) {
+                            const newAllComments = [...allComments];
+                            newAllComments.splice(i, 1);
+                            setAllComments(newAllComments);
+                          }
+                        }}
+                      />
+                    );
+                  }
                 })}
                 <div className="flex w-full mt-8">
                   <div className="flex-none mr-4">
@@ -233,7 +245,9 @@ function RepositoryIssueView(props) {
                         <img
                           src={
                             "https://avatar.oxro.io/avatar.svg?length=1&height=100&width=100&fontSize=52&caps=1&name=" +
-                            (props.activeWallet ? props.activeWallet.name : "")
+                            (props.selectedAddress
+                              ? props.selectedAddress.slice(-1)
+                              : "")
                           }
                         />
                       </div>
@@ -253,46 +267,89 @@ function RepositoryIssueView(props) {
                   assignees={issue.assignees}
                   collaborators={repository.collaborators}
                   onChange={async (list) => {
-                    const res = await props.updateIssue({
+                    console.log("list", list);
+                    const removedAssignees = issue.assignees.filter(
+                      (x) => !list.includes(x)
+                    );
+                    const addedAssignees = list.filter(
+                      (x) =>
+                        !(
+                          removedAssignees.includes(x) ||
+                          issue.assignees.includes(x)
+                        )
+                    );
+
+                    const res = await props.updateIssueAssignees({
                       issueId: issue.id,
-                      assignees: list,
+                      addedAssignees,
+                      removedAssignees,
                     });
 
                     if (res) refreshIssue();
                   }}
                 />
-                <div className="text-xs px-3">
+                <div className="text-xs px-3 mt-2">
                   {issue.assignees.length
-                    ? issue.assignees.map((a) => shrinkAddress(a)).join(", ")
+                    ? issue.assignees.map((a, i) => (
+                        <span
+                          className="pr-2 pb-2 whitespace-nowrap"
+                          key={"assignee" + i}
+                        >
+                          <a
+                            href={"/" + a}
+                            className="btn-link cursor-pointer"
+                            target="_blank"
+                          >
+                            {shrinkAddress(a)}
+                          </a>
+                        </span>
+                      ))
                     : "No one"}
                 </div>
               </div>
               <div className="py-8">
                 <LabelSelector
                   labels={issue.labels}
+                  repoLabels={repository.labels}
+                  editLabels={
+                    "/" +
+                    repository.owner.id +
+                    "/" +
+                    repository.name +
+                    "/issues/labels"
+                  }
                   onChange={async (list) => {
-                    const res = await props.updateIssue({
+                    console.log("list", list);
+                    const removedLabels = issue.labels.filter(
+                      (x) => !list.includes(x)
+                    );
+                    const addedLabels = list.filter(
+                      (x) =>
+                        !(removedLabels.includes(x) || issue.labels.includes(x))
+                    );
+
+                    const res = await props.updateIssueLabels({
                       issueId: issue.id,
-                      labels: list,
+                      addedLabels,
+                      removedLabels,
                     });
 
                     if (res) refreshIssue();
                   }}
                 />
-                <div className="text-xs px-3">
+                <div className="text-xs px-3 mt-2 flex flex-wrap">
                   {issue.labels.length
                     ? issue.labels.map((l, i) => {
-                        let label = allLabels[l] ?? {
+                        let label = _.find(allLabels, { id: l }) || {
                           name: "",
                           color: "",
                         };
                         return (
                           <span
-                            className="badge badge-sm p-2 border-0 mr-2 mb-2"
-                            style={{ backgroundColor: label.color }}
-                            key={"issueLabel" + i}
+                            className="pr-2 pb-2 whitespace-nowrap"
+                            key={"label" + i}
                           >
-                            {label.name}
+                            <Label color={label.color} name={label.name} />
                           </span>
                         );
                       })
@@ -300,11 +357,11 @@ function RepositoryIssueView(props) {
                 </div>
               </div>
               <div className="py-8">
-                <div className="flex-1 text-left  px-3 mb-1">
+                <div className="flex-1 text-left px-3 mb-1">
                   Linked Pull Requests
                 </div>
 
-                <div className="text-xs px-3">None yet</div>
+                <div className="text-xs px-3 mt-2">None yet</div>
               </div>
             </div>
           </div>
@@ -322,6 +379,8 @@ const mapStateToProps = (state) => {
   };
 };
 
-export default connect(mapStateToProps, { deleteComment, updateIssue })(
-  RepositoryIssueView
-);
+export default connect(mapStateToProps, {
+  deleteComment,
+  updateIssueAssignees,
+  updateIssueLabels,
+})(RepositoryIssueView);
